帖子详情页面删除，先确认

+++ src/pages/forum/previewDetails.vue
@@ -180,13 +180,24 @@ export default {
   },
   methods: {
     deleteArticle () {
-      deleteArticle({ ids: arrayToStrWithOutComma(this.article.id.split(',')) }).then(res => {
-        if (!res) {
-          this.$message.error('删除失败，请稍后再试')
-        } else {
-          this.$message.success('删除成功')
-          this.$router.push({ name: 'home' })
-        }
+      this.$confirm('此操作将删除该帖子, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteArticle({ ids: arrayToStrWithOutComma(this.article.id.split(',')) }).then(res => {
+          if (!res) {
+            this.$message.error('删除失败，请稍后再试')
+          } else {
+            this.$message.success('删除成功')
+            this.$router.push({ name: 'home' })
+          }
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
     },
     goBacktToHome () {
